Unifying Problem interface
