+++ test/validator.test.ts
@@ -1,9 +1,9 @@
-import { Node } from './../src/model'
+import { Node, Problem } from './../src/model'
 import { Annotation, buildEnvironment } from '../src'
 import globby from 'globby'
 import { readFileSync } from 'fs'
 import { join } from 'path'
-import validate, { Problem } from '../src/validator'
+import validate from '../src/validator'
 import { should } from 'chai'
 import { fail } from 'assert'
 import { notEmpty } from '../src/extensions'
